# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Enhanced RAG Chatbot API",
    "version": "2.0.0",
    "description": "Comprehensive API documentation for the Enhanced RAG Chatbot with JWT Authentication, Dual Database Architecture, and Free AI Components",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    },
    {
      "url": "https://api.yourapp.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication. Format: 'Bearer <token>'"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error type"
          },
          "message": {
            "type": "string",
            "description": "Detailed error message"
          },
          "code": {
            "type": "integer",
            "description": "Error code"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "paths": {
...
    "/chats/{chatId}/upload": {
      "post": {
        "summary": "Upload PDF document to a specific chat",
        "tags": ["Chat Management"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "chatId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the chat to upload document to"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "PDF file to upload"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document uploaded and processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "document": {
                      "type": "object",
                      "properties": {
                        "filename": {
                          "type": "string"
                        },
                        "id": {
                          "type": "string"
                        },
                        "textLength": {
                          "type": "integer"
                        },
                        "pages": {
                          "type": "integer"
                        },
                        "processingResults": {
                          "type": "object",
                          "properties": {
                            "databaseStored": {
                              "type": "boolean"
                            },
                            "vectorStored": {
                              "type": "boolean"
                            },
                            "ragEnabled": {
                              "type": "boolean"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid file format or upload error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/chats/{chatId}/documents": {
      "get": {
        "summary": "Get list of documents in a specific chat",
        "tags": ["Chat Management"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "chatId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID of the chat"
          }
        ],
        "responses": {
          "200": {
            "description": "List of documents in the chat",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "chatId": {
                      "type": "string"
                    },
                    "documents": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "filename": {
                            "type": "string"
                          },
                          "uploadedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "size": {
                            "type": "integer"
                          },
                          "pages": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}